requirements.txt
torch
torchvision
opencv-python
Pillow
PyYAML
mlflow
ultralytics
supervision